Optionally return the full response
The OpenAI response contains useful information in the
headers (notably this is where the [rate limiting
info](https://platform.openai.com/docs/guides/rate-limits#rate-limits-in-headers) is returned).

This change allows the user to specify that they want the whole
response returned instead of just the body, by setting the
`:full-response?` option.

This change should be backwards compatible and so shouldn't break
existing code.